move registry, begin docs
