+++ adept/scripts/impala.py
@@ -25,7 +25,7 @@ from tensorboardX import SummaryWriter
 
 from adept.containers import ImpalaHost, ImpalaWorker
 from adept.environments import SubProcEnvManager, EnvMetaData
-from adept.environments.registry import EnvPluginRegistry
+from adept.registries.environment import EnvPluginRegistry
 from adept.utils.logging import (
     make_log_id_from_timestamp, make_logger,
     print_ascii_logo, log_args, write_args_file, SimpleModelSaver
