+++ adept/scripts/replay_gen_sc2.py
@@ -23,7 +23,7 @@ from absl import flags
 
 from adept.containers import ReplayGenerator
 from adept.environments import SimpleEnvManager
-from adept.environments.registry import EnvPluginRegistry, Engines
+from adept.registries.environment import EnvPluginRegistry, Engines
 from adept.utils.logging import print_ascii_logo
 from adept.utils.script_helpers import make_agent, make_network, \
     get_head_shapes, parse_bool
